whatchamp/requirements.txt
flask
numpy
pandas
requests
scikit-learn
flask-cors
transformers
python-dotenv
openpyxl
